project-jobshop_alifassyfehry-abdellahsofi/wordle-solver/backend/requirements.txt
# API Backend
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
google-genai>=0.2.0
